services/vanna/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.28.0
vanna==0.5.5
python-dotenv==1.0.1
psycopg2-binary==2.9.9
groq==0.4.2
pydantic==2.6.3
